fix: crash because `mergedMessage` was 'nil' in `strings.SplitAfter()`
It should never be nil. This was happening because when `firstIdx` was the
previous to the last one, and they had the same start location, we left early
the while loop without merging the lines collected.

Example: in a list of 34 messages, if we were at 33 and 33 and 34 had the same
location, then we left the while loop early without merging lines into
`mergedMessage`.